Protocol: Implemented GenKey command. Implemented Nonce command Implemented Sign command implemented Random command. Command execution will now ignore the result of the wake sequence. Command execution will now poll for command result for at most 500 milliseconds.
Driver: Implemented ReadSerialNumber.
Implemented GetPublicKey API.
Implemented Nonce API.
Implemented Random API.
Implemented Sign API.
Added mutex to all public methods of Device to prevent race condition when communicating with the secure element.

+++ constants.go
@@ -558,7 +558,7 @@ const (
 	RANDOM_COUNT                = ATCA_CMD_SIZE_MIN //!< Random command packet size
 	RANDOM_SEED_UPDATE    uint8 = 0x00              //!< Random mode for automatic seed update
 	RANDOM_NO_SEED_UPDATE uint8 = 0x01              //!< Random mode for no seed update
-	RANDOM_NUM_SIZE       uint8 = 32                //!< Number of bytes in the data packet of a random command
+	RANDOM_NUM_SIZE       int   = 32                //!< Number of bytes in the data packet of a random command
 	RANDOM_RSP_SIZE             = ATCA_RSP_SIZE_32  //!< Random command response packet size
 	/** @} */
